Add files via upload
I added fluorescence at the cast_to_xarray.py

I also added fluorescence plots at the end of the codes

In your code I had to make some small modifications on the name of the xarray files. I did this because you made xarray files with the name: "castxr6", but when you called it to concatenate, you called "cast6xr". So I had to modify one of these.